% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/biocontainer.r
\name{biocontainer}
\alias{biocontainer}
\title{biocontainer}
\usage{
biocontainer(type = "Trap  REDLAV ITA", nrecipients = 50,
  container_shape = 3, capacity = 1000, initVol = 750,
  frac_initVol = 0.75, surface_area = 50, underplate = FALSE,
  frac_underplate = 0, frac_covering = 0, lat = 43.5, lon = 11.27,
  elevation = 100, CRS = "+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs",
  sourcedata = NULL, meteodata = NULL, watermodel = NULL,
  watermodel_fit = FALSE, model_type = "gam", date_format = "YMD",
  field_delimiter = ",", ID = c("NA"), site_name = c("NA"))
}
\arguments{
\item{type}{char string:   Container type.  Default Trap used in REDLAV project: volume 0.75 750 cm^3 lt, heigth 13 cm, Area 50 cm^2.Describe the tipology of cointainer considered. For example (i) cointaner (ii) manhole (iii) others.}

\item{nrecipients}{numeric:  Numerosity of reply simulating the number of breeding sites BS. Default ( 1 lt 1000 cm^3).}

\item{container_shape}{numeric: If a surface container a numerical label indicate the shape . Parallelepiped plate (1)  Cilynder plate(2) Unspecified (3). Default is 3.}

\item{capacity}{numeric:  Maximum capacity of water in trap/container. Default ( 1 lt 1000 cm^3).}

\item{frac_initVol}{numeric: Fraction of volume present in trap/location. Default is 0.}

\item{surface_area}{numeric: water-air surface of container in squared centimeter (cmq).}

\item{underplate}{logical: Presence of underplate. Default is FALSE.}

\item{frac_underplate}{numeric: Fraction of underplate engaged ( 0-1) Default is 0.}

\item{frac_covering}{numeric: Fraction of container covering  ( 0-1) Default is 0.}

\item{lat}{numeric: Mean geographical latitude of container location. Default is 43.5.}

\item{lon}{numeric: Mean geographical longitude of container location. Default is 11.27.}

\item{elevation}{numeric:  Mean elevation  of container location. Default is 100.}

\item{CRS}{char string: Projection of coordinate in proj4 format.}

\item{sourcedata}{data.frame or ascii file:  Water temperature measurements. Fields required are dates ( YYYY-MM-DD) format, Daily mean of water tmedwater.}

\item{meteodata}{Rbiosim object:  with the data useful for model water calibration.}

\item{watermodel}{logical: Statistical model oject ( gam - lm - glm).Default is NULL.}

\item{watermodel_fit}{logical: Fit a statistical model oject ( gam - lm - glm).Default is FALSE.}

\item{model_type}{char string:  Statistical model class.}

\item{date_format}{char string: Format of the dates in raw data. Default is YMD.}

\item{field_delimiter}{char string: field delimiter of file. Default is comma ",".}

\item{ID}{char string: ID of container's set.}

\item{init_Vol}{numeric:  Unitarian initial volume of water in trap/container. Default ( 0.75 lt 750 ml).}

\item{daily_evap_loss}{numeric: Volume of water losses daily due to evaporation. Default is 0.}

\item{timeseries}{xts R object: Timeseries of data}
}
\value{
Rbiosim object: Biocontainer object.
}
\description{
\code{biocontainer} S3 class for container object to be used in rAedesSim model.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso crisci \email{a.crisci@ibimet.cnr.it} ASL 2 LUCCA Marco Selmi \email{m.selmi@usl2.toscana.it}
}
\keyword{container}

