% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/biodata.r
\name{biodata}
\alias{biodata}
\title{biodata}
\usage{
biodata(parameter_definition = "Egg counts in trap",
  unity_measure = "Integer count", location,
  instrument = "Trappola REDLAV Ita", network = "",
  common_name = "Zanzara Tigre", scientific_name = "Aedes Albopictus",
  phenology = "Eggs Hatching", obs_standard = "", sourcedata,
  data_provider = "ASL Lucca",
  data_maintaner = "Marco Selmi - m.selmi@usl2.toscana.it",
  data_licence = "", field_delimiter = ",", lat, lon, elevation,
  CRS = "+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs", geonotes = "",
  ID, urban, nasa_radiance, feature_population)
}
\arguments{
\item{parameter_definition}{char string: Name of site of observations.}

\item{location}{char string: Name of site of observations.}

\item{instrument}{char string: Name of engine to provide observation}

\item{network}{char string: If data is collected inside an observative network.}

\item{common_name}{char string: Data tipology - Observation - Sensor monitoring - Simulation}

\item{scientific_name}{char string: Data tipology - Observation - Sensor monitoring - Simulation}

\item{phenology}{char string: Data tipology - Observation - Sensor monitoring - Simulation}

\item{obs_standard}{char string: if data belong to a phenology Data standard class es SYNOP}

\item{sourcedata}{Matrix or data.frame or ascii file of raw data.}

\item{data_provider}{char string: Institution / Private data manager.}

\item{data_maintaner}{char string: maintainer's  name or contact or its contact .}

\item{data_licence}{char string: Licence of data.}

\item{field_delimiter}{char string: field delimiter of file. Default is comma ",".}

\item{lat}{numeric: latitude coordinates of  where data are collected.}

\item{lon}{numeric: longitude coordinates of  where data are collected.}

\item{elevation}{numeric: Elevation of the Default is 40.}

\item{CRS}{char string: Projection of coordinate in proj4 format.}

\item{geonotes}{char string: Annotations in regard to the contest of observation.}

\item{ID}{char string: ID of monitoring.}

\item{urban}{logical: Flag indicating if data belong to urban area. Automatic detection is done in according to UMZ Urban Morfological Zone EAA http://database.espon.eu}

\item{nasa_radiance}{numeric: Night radiance value. Is a proxy for urbanity.}

\item{units}{char string: Measure unity of observations.}

\item{sourcedata}{Matrix or data.frame or ascii file of raw data. Field required are dates and parameter.}

\item{CRS}{char string: Projection of coordinate in proj4 format.}

\item{feature_population:}{population density estimation.}
}
\value{
Rbiosim object Biodata object
}
\description{
\code{biodata} Biodata is a rAedesSim S3 class to describe data collection from the field concerning.
}
\author{
Istituto di Biometeorologia Firenze Alfonso crisci \email{a.crisci@ibimet.cnr.it} ASL 2 LUCCA  Marco Selmi \email{m.selmi@usl2.toscana.it}
}
\keyword{biomodel}

