% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/biometeo.r
\name{biometeo}
\alias{biometeo}
\title{biometeo}
\usage{
biometeo(meteo, i_biocontainer, deltatmax = 0, deltatmin = 0,
  deltatmed = 0, tresh_rain = 4, weigth_k = 5)
}
\arguments{
\item{deltatmax}{numeric: Mean Error Bias  considered for maximum air temperature. Default is 0.}

\item{deltatmin}{numeric: Mean Error Bias  considered for minimum air temperature. Default is 0.}

\item{deltatmed}{numeric: Mean Error Bias  considered for mean air temperature. Default is 0.}

\item{weigth_k}{Weighting Paraemter to take into account dry spell impact See also \code{\link{weigthdry}}. Default is 7.}

\item{meteodata}{Rbiosim object.}

\item{biocontainer}{Rbiosim object .}

\item{tresh_rain_dry}{numeric: Rain threshold for effective precipitation. Default is 4.}
}
\value{
Return a biometeo object.
}
\description{
S3
\code{biometeo} class to describe  object to be used in Rbiosim model.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso crisci \email{a.crisci@ibimet.cnr.it} ASL 2 LUCCA Marco Selmi \email{m.selmi@usl2.toscana.it}
}
\keyword{container}

