% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/biomodel.r
\name{biomodel}
\alias{biomodel}
\title{biomodel}
\usage{
biomodel(i_biopopulation, i_bioparameters, i_biometeo, i_biocontainer,
  state = c(L1 = 0, L3 = 0, L4 = 0, L5 = 0, L6 = 0, L7 = 0, L8 = 0, L10 = 0),
  time_interval = 24, stocastic = TRUE, n_sampling = 10,
  inibition = FALSE, ID_sim = "")
}
\arguments{
\item{i_biopopulation}{object: rAedesSim object concerning mosquito population.}

\item{i_bioparameters}{object:rAedesSim object concerning biological parameter of mosquito population.}

\item{i_biometeo}{object:rAedesSim object concerning biometeorological parameter of location where mosquito population are simulated.}

\item{i_biocontainer}{object:rAedesSim object concerning trap or habitat considered.}

\item{state}{numeric: State vector  for simulation.The state considered are L1 : Ovideposition rate, L3 : Eggs Mortality rate,
L4: Eggs2larvae transition rate, L5 : Pupae2Larvae transition rate, L6 larvae Mortality,L7  pupae Mortality
L8 Pupae2adult transition rate, L10 Female adult mortality Default : L1=0,L3=0,L4=0,L5=0,L6=0,L7=0,L8=0,L10=0}

\item{time_interval}{numeric: Interval times of simulation.}

\item{stocastic}{logical: performs stocastic elaboration by poisson sampling at the end of simulation for each state.}

\item{inibition}{logical: taking into account density larvae inibition.Default is FALSE.}

\item{ID_sim}{char string: ID of simulation.Default is FALSE.}

\item{n_sampling=10}{numeric : number of resampling.}
}
\value{
Rbiosim object Biomodel object
}
\description{
\code{biomodel} S3 class for biomodel object to be used in rAedesSim package. This is the object that lead simulation.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso crisci \email{a.crisci@ibimet.cnr.it} ASL 2 LUCCA Marco Selmi \email{m.selmi@usl2.toscana.it}
}
\keyword{biomodel}

