% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bioparameters.r
\name{bioparameters}
\alias{bioparameters}
\title{bioparameters}
\usage{
bioparameters(alfa_l = 1.5, alfa_a = 0, exp_lar = 1, l_density = 70,
  sex_ratio = 0.5, egn = 63, ef = 0.83, inib = 0.63,
  sspp = "Albopictus", genus_sspp = "Aedes", order_sspp = "Diptera",
  geo_area = "Tuscany", name_location = NA)
}
\arguments{
\item{Numeric}{alfa_l Coefficient of competition between individuals at larval stage.}

\item{Numeric}{alfa_a  Coefficient of competition between individuals at adult stage.}

\item{Numeric}{exp_lar Exponential Gompertz parameter for competition modulation. Default=1.}

\item{Numeric}{l_density Expected larval density. Default is 70.}

\item{Numeric}{sexratio   Sex ratio of population. Default is 0.5.}

\item{Numeric}{egn  Mean eggs . Default is 63.}

\item{Numeric}{inib Inibition rate parameter ( 0-1) Default is 0.}

\item{String}{sspp Name of the species. Default is "Albopictus".}

\item{String}{genus_sspp  Name of the genus of the species. Default is "Aedes".}

\item{String}{order_sspp  Name of the order of the species. Default is "Diptera".}

\item{String}{geo_area  Name of geographic area. Default is "Tuscany".}

\item{String}{name_location  Name of location.}

\item{Numeric}{name_location Mean elevation  of container location. Default is 100.}
}
\value{
S3 object Bioparameters object
}
\description{
\code{ bioparameters} This is a S3 class describing biological parameters used in simulation for rAedesSim .
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso crisci \email{a.crisci@ibimet.cnr.it} ASL 2 LUCCA Marco Selmi \email{m.selmi@usl2.toscana.it}
}
\keyword{biparameters}

