% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/biopopulation.r
\name{biopopulation}
\alias{biopopulation}
\title{biopopulation}
\usage{
biopopulation(eggs = 10000, larvae = 0, pupae = 0, adults = 0,
  eggs_diap = 10, ID = "NA")
}
\arguments{
\item{eggs}{Numeric Eggs Count.}

\item{larvae}{Numeric Larvae Count.}

\item{pupae}{Numeric Pupae Count.}

\item{adults}{Numeric Mosquito Female Actual Count.}

\item{eggs_diap}{Numeric Diapausant Eggs Count.}

\item{ID}{String Population.}
}
\value{
S3 object biopopulation object.
Istituto di Biometeorologia Firenze Italy
ASL 2 LUCCA
}
\description{
S3 class to describe biomodelday object to be used in rAedesSim model.
}
\author{
Alfonso crisci \email{a.crisci@ibimet.cnr.it} Marco Selmi \email{m.selmi@usl2.toscana.it}
}
\keyword{container}

