% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_fit_initial.r
\name{check_fit_initial}
\alias{check_fit_initial}
\title{check_fit_initial}
\usage{
check_fit_initial(i_biometeo, i_monitoring, i_biocontainer, larv_alpha,
  adu_alpha, eggs_diap, Ndayscenario_prim = 120, Ndayscenario_aut = 90,
  deltatmed_prim = 0, deltawater_prim = 0, deltatmed_aut = 0,
  deltawater_aut = 0, ini_rmse = 1, end_rmse = NULL, stocastic = FALSE,
  n_sampling = 10, saveparameter = TRUE)
}
\arguments{
\item{i_biometeo}{object rAedesSim \code{ biometeo} object}

\item{i_biocontainer}{object rAedesSim \code{biocontainer} object .}

\item{larv_alpha}{numeric alpha larvae intra-competition factor.}

\item{adu_alpha}{numeric adult larvae intra-competition factor.}

\item{eggs_diap}{numeric Number of diapausant eggs considered.}

\item{Ndayscenario_prim}{numeric Indicate the number of days considered after the first date of meteo data series. Defalut is 120.}

\item{Ndayscenario_aut}{numericIndicate the number of days considered before the last date of meteo data series. Defalut is 90.}

\item{deltatmed_prim}{numeric Mean air temperature bias cosidered for spring period.}

\item{deltawater_prim}{numeric Mean water temperature bias cosidered for spring period.}

\item{deltatmed_aut}{numeric Mean air temperature bias cosidered for spring period.}

\item{deltawater_aut}{numeric Mean water temperature bias cosidered for spring period.}

\item{ini_rmse}{numeric Starting index to calculate RMSE. Defalut is 1 that means all monitoring data are considered.}

\item{end_rmse}{numeric Ending index to calculate RMSE. Defalut is NULL that means all monitoring data are considered.}

\item{stocastic}{logical performs stocastic elaboration by poisson sampling at the end of simulation for each state.}

\item{n_sampling}{numeric number  of resampling if stochastic is implemented see in \code{biomodel}. Default is 10.}

\item{saveparameter}{logical  Save results about parameters. Default is FALSE.}

\item{i_moitoring}{object rAedesSim \code{biodata} object concerning mosquito eggs field observations.}
}
\value{
List of results with graph, esclosion graph ,the initial rmse and the full simulation object.
}
\description{
check_fit_initial is a function to identificate in a unknow location 
which are the best initial thermal condition partial to fit better the start of season and in spring and autmn the thermal bias due to  air and water temperature.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso crisci \email{a.crisci@ibimet.cnr.it} ASL 2 LUCCA Marco Selmi \email{m.selmi@usl2.toscana.it}
}
\keyword{biofitmodel}
