% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cmnl.r
\name{cmnl}
\alias{cmnl}
\title{cmnl}
\usage{
cmnl(alfa0, nrecipients)
}
\arguments{
\item{alfa0}{Number Is the  competition parameter at larval stage.}

\item{nrecipients}{Number Number of container defined in simulation.}
}
\value{
Numeric proxy of carrying capacity saying how many container are engaged in mosquito spreading.
}
\description{
Calculate a proxy of carryng capacity as ratio between alfa0 larval factor and breeding site (BS) number
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso crisci \email{a.crisci@ibimet.cnr.it} ASL 2 LUCCA Marco Selmi \email{m.selmi@usl2.toscana.it}
}
\references{
Otero, M., Solari, H., Schweigmann, N., 2006. A stochastic population dynamic model for aedes aegypti: formulation and application to a city with temperate climate. Bull. Math. Biol. 68, 1945–974.
}
\keyword{Container}

