% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/container_capacity.r
\name{container_capacity}
\alias{container_capacity}
\title{container_capacity}
\usage{
container_capacity(base_area, heigth, underplate = FALSE, fraction_underplate)
}
\arguments{
\item{base_area}{Basal area of container ( square meters).}

\item{heigth}{Heigth of container ( meters) .}

\item{underplate}{Presence of underplate Default underplate=FALSE}

\item{fraction_underplate}{Fraction of undeplate engaged}
}
\value{
Volume capacity of container in liters
}
\description{
Calculate the volume of cointaner taking into account the underplate precence.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso crisci \email{a.crisci@ibimet.cnr.it} ASL 2 LUCCA Marco Selmi \email{m.selmi@usl2.toscana.it}
}
\keyword{Container}

