% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/day_length.r
\name{day_length}
\alias{day_length}
\title{day_length}
\usage{
day_length(dates, lon = 11.1, lat = 43.77)
}
\arguments{
\item{dates}{Date in  YMD format}

\item{lon}{Numerical Geographical Longitude of site in decimal degrees (negative == West) . Default is 11.10.}

\item{lat}{Numeric Geographical Latitude of site in decimal degrees. Default is 43.77.}
}
\value{
Return sunrise time, suset time,length of day in hours decimal.
}
\description{
Calculate the raw lenght of day in hours related to the date
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso crisci \email{a.crisci@ibimet.cnr.it} ASL 2 LUCCA Marco Selmi \email{m.selmi@usl2.toscana.it}
}
\references{
Teets, D.A. 2003. Predicting sunrise and sunset times. The College Mathematics Journal 34(4):317-321.
}
\seealso{
\code{\link{},\link{weigthdry}}
}
\keyword{days}
\keyword{dry}

