% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/psychro_function.r
\name{degsat}
\alias{degsat}
\title{degsat}
\usage{
degsat(tair = 24, rh = 50)
}
\arguments{
\item{tair}{numeric: air temperature in Celsius}

\item{urel}{numeric: Air Relative umidity (\%)}
}
\value{
numeric: saturation deficit pressure in millibar 10-30.
}
\description{
Air degree of saturation. Is a meausere of air dryness
}

