% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/psychro_function.r
\name{delta_vap_pressure}
\alias{delta_vap_pressure}
\title{delta_vapor_pressure}
\usage{
delta_vap_pressure(tair = 24, twater = 21, urel = 60)
}
\arguments{
\item{tair}{numeric: Air temperature in Celsius}

\item{twater}{numeric: Water temperature in Celsius}

\item{urel}{numeric: Air Relative umidity (\%)}
}
\value{
vap_pressure numeric: Vapour pressure gradient
}
\description{
Difference water vapour pressure between cointaner water surface an free air.
}

