% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/psychro_function.r
\name{delta_vapor_density}
\alias{delta_vapor_density}
\title{delta_vapor_density}
\usage{
delta_vapor_density(tair = 24, twater = 21, urel = 60)
}
\arguments{
\item{tair}{numeric: air temperature in Celsius}

\item{twater}{numeric: water temperature in Celsius}

\item{urel}{numeric: realtive humidity}
}
\value{
delta_vap_interface numeric: gradient of vapour pressure at surface air-water
}
\description{
Saturation pressure of water in air as function of temperature (°C)
}

