% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/develop_rate.r
\name{develop_rate}
\alias{develop_rate}
\title{develop_rate}
\usage{
develop_rate(temp, DR25, EA, EI, TI, t25 = 298)
}
\arguments{
\item{temp}{Numeric  Temperature in Celsius degree.}

\item{DR25}{Numeric Developmental rate standard at 25 Celsius.}

\item{EA}{Numeric Enthalpy of activation for determinate stage.}

\item{EI}{Numeric Enthalpy of inactivation for determinate stage.}

\item{TI}{Numeric Temperature of inactivation for determinate stage.}

\item{t25}{Numeric Temperature of inactivation for determinate stage.}
}
\value{
Numeric Develompental rate for
}
\description{
Calculate developmental rates for Aedes in function to the air temperature.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso crisci \email{a.crisci@ibimet.cnr.it} ASL 2 LUCCA Marco Selmi \email{m.selmi@usl2.toscana.it}
}
\references{
Otero, M., Solari, H., Schweigmann, N., 2006. A stochastic population dynamic model for aedes aegypti: formulation and application to a city with temperate climate. Bull. Math. Biol. 68, 1945–974.
}
\keyword{develpmentantal}
\keyword{rate}

