% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/psychro_function.r
\name{dewpoint}
\alias{dewpoint}
\title{dewpoint}
\usage{
dewpoint(tair = 24, urel = 50)
}
\arguments{
\item{t}{numeric: air temperature in Celsius}
}
\value{
numeric: Return dewpoint in Celsius
}
\description{
Saturation pressure of water in air as function of temperature (°C)

The saturation pressure of water is the pressure at or above which water
would condense from from the air. It is a function of temperature only.
}

