% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diapause_emergency.r
\name{diapause_emergency}
\alias{diapause_emergency}
\title{diapause_emergency}
\usage{
diapause_emergency(dates, lon = 7.216, lat = 43.658, CPPmin = 11.5,
  jdmax = 110, bounddays = 10, varjd = 10, truncated = F)
}
\arguments{
\item{dates}{Date in  YMD format}

\item{lon}{Numeric Geographical Longitude of site in decimal degrees. Default is 7.216.}

\item{lat}{Numeric Geographical Latitude of site in decimal degrees. Default is 43.77.}

\item{CPPmin}{Numeric Number of hours of critical photoperiod. Default is 11.5}

\item{bounddays}{Numeric Days of incertain for diapause emergency bounds. Default is 10.}

\item{varjd}{Numeric Days of variance for  diapause emergency courbes in days. Default is 10.}

\item{truncated}{logical Modeling with truncated gaussian respecting the limits. Default is FALSE.}

\item{datemax}{Date Date of spring diapause awekaning in MM-DD. Default is 04-20.}
}
\value{
Return diapause emergency in percentage.
}
\description{
Calculate the daily diapause incidence of egg springtime hatching in probabilistic terms.
}
\references{
Lacour G, Chanaud L, L’Ambert G, Hance T (2015) Seasonal Synchronization of Diapause Phases in Aedes albopictus Diptera: Culicidae. PLoS ONE 10-12: e0145311. doi:10.1371 journal.pone.0145311
}
\seealso{
\code{diapause_induction}
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso crisci \email{a.crisci@ibimet.cnr.it} ASL 2 LUCCA Marco Selmi \email{m.selmi@usl2.toscana.it}
}
\keyword{,spring}
\keyword{diapause}
\keyword{emergency}
