% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diapause_induction.r
\name{diapause_induction}
\alias{diapause_induction}
\title{diapause_induction}
\usage{
diapause_induction(dates, lon = 7.216, lat = 43.658)
}
\arguments{
\item{dates}{Date in  YMD format}

\item{lon}{Numeric Geographical Longitude of site in decimal degrees. Default is 7.216.}

\item{lat}{Numeric Geographical Latitude of site in decimal degrees. Default is 43.77.}
}
\value{
Return Pourcentage of diapausant eggs.
}
\description{
Calculate the daily diapause incidence  in autumnal eggs of Aedes albopictus. 
It is based on the work published by Lacour G. at al (2015), Seasonal Synchronization of Diapause Phases in Aedes albopictus (Diptera: Culicidae),
in PLOS. Date limits are imposed by default. Julian dates 237 for diapause starting and 357 for diapause ending. It needs location coordinates where throughout  \code{"day_length"} function astronomical daily sunlight duration is calculated.
}
\references{
Lacour G, Chanaud L, L’Ambert G, Hance T (2015) Seasonal Synchronization of Diapause Phases in Aedes albopictus (Diptera: Culicidae). PLoS ONE 10 12 : e0145311. doi:10.1371 journal.pone.0145311
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso crisci \email{a.crisci@ibimet.cnr.it} ASL 2 LUCCA Marco Selmi \email{m.selmi@usl2.toscana.it}
}
\keyword{diapause,autumn}
