% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diapause_model.r
\docType{data}
\name{diapause_model}
\alias{diapause_model}
\title{diapause_model}
\format{A lm R object with polinomial indidependent of daily length ( degree=8).}
\source{
\url{http://www.ibimet.cnr.it/}
}
\usage{
diapause_model
}
\description{
Numerical function to assess diapause induction in eggs of Aedes albopictus.It is based on Lacour G. at al (2015) work.
}
\references{
Lacour G, Chanaud L, L’Ambert G, Hance T (2015) Seasonal Synchronization of Diapause Phases in Aedes albopictus (Diptera: Culicidae). PLoS ONE 10(12): e0145311. doi:10.1371/journal.pone.0145311
}
\keyword{datasets}
