% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/diapause_rate.r
\name{diapause_rate}
\alias{diapause_rate}
\title{diapause_rate}
\usage{
diapause_rate(daylength)
}
\arguments{
\item{dates}{Date in  YYYY-MM-DD format}

\item{lon}{Numerical Geographical Longitude of site. Default is 11.10.}

\item{lat}{Numeric Geographical Latitude of site. Default is 43.77.}
}
\value{
Rate of egg exclosion and inversaly the diapausant one.
}
\description{
Calculate the percentage of emergency of diapausant eggs in function to legth of day.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso crisci \email{a.crisci@ibimet.cnr.it} ASL 2 LUCCA Marco Selmi \email{m.selmi@usl2.toscana.it}
}
\seealso{
\code{\link{},\link{trans_rate}}
}
\keyword{diapause}

