% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/psychro_function.r
\name{evaporation_params}
\alias{evaporation_params}
\title{evaporation_params}
\usage{
evaporation_params(tair = 24, twater = 21, urel = 60, v_air = 0.5,
  L = 0.1, A = NULL)
}
\arguments{
\item{tair}{numeric: Air temperature in Celsius}

\item{twater}{numeric: Water temperature in Celsius}

\item{urel}{numeric: Air Relative umidity (\%)}

\item{v_air}{numeric: Velocity of air movement (m/sec)}

\item{L}{numeric: Dimension caracterize the cointainer. Generally the diameter (m)}
}
\value{
evap_pars numeric list:  Element mevday give a the mass of  evaporate (mg) for day for a cointanertaking into account dimension.
}
\description{
Give a set of parameters  to assess the evaporative losses for a generic cointainer under forced convetion (v_air > 0.1).
}

