% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/psychro_function.r
\name{mass_diffusion}
\alias{mass_diffusion}
\title{mass_diffusion_vap}
\usage{
mass_diffusion(tair = 24)
}
\arguments{
\item{tair}{numeric: Air temperature in Celsius}
}
\value{
m_vap numeric: Diffusion coefficient of vapor (mq/s).
}
\description{
Diffusivity or diffusion coefficient of vapor in function to  temperature for Air-Water Vapor Mixtures  Bolz and Tuve (1976).
}

