% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychro_function.r
\name{mass_diffusion_vap}
\alias{mass_diffusion_vap}
\title{mass_diffusion_vap}
\usage{
mass_diffusion_vap(tair = 24)
}
\arguments{
\item{tair}{numeric: Air temperature in degC.}
}
\value{
mass_diffusion_vap
}
\description{
@description  Diffusion coeffcient of vapor (mq/s).in function to  temperature for Air-Water Vapor Mixtures  Bolz and Tuve (1976).
}
