% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/psychro_function.r
\name{mass_transport_coef}
\alias{mass_transport_coef}
\title{mass_transport_coef}
\usage{
mass_transport_coef(tair = 23, twater = 21, urel = 50, v_air = 0.5)
}
\arguments{
\item{tair}{numeric: Air temperature in Celsius}

\item{twater}{numeric: Water temperature in Celsius}

\item{urel}{numeric: Air Relative umidity (\%)}

\item{v_air}{numeric: Velocity of air movement (m/sec)}
}
\value{
numeric: Mass transfer coefficient for water. Only vapour pressure gradient is considered.
}
\description{
Mass transfer coefficient in function to air temperature in forced convection  (v_air > 0.1).
}

