% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/meteodata.r
\name{meteodata}
\alias{meteodata}
\title{meteodata}
\usage{
meteodata(station_name = "Pisa San Giusto",
  network = "Aeronautica Militare", data_type = "Simulation",
  standard = "SYNOP", data_provider = "IBIMET CNR", data_maintainer = "",
  data_licence = "", date_format = "YMD", lat = 43, lon = 11,
  elevation = 40, timeformat = "daily", sourcedata = NULL,
  field_delimiter = ",", timeseries = NULL,
  CRS = "+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs")
}
\arguments{
\item{station_name}{char string: Name of station}

\item{network}{char string: Observative or set of data}

\item{data_type}{char string: Data tipology - Observation - Sensor monitoring - Simulation}

\item{standard}{char string: Data standard class es SYNOP}

\item{data_provider}{char string: Institution / Private data manager.}

\item{data_licence}{char string: Licence of data.}

\item{date_format}{char string: Format of the dates in raw data. Default is YMD.}

\item{lat}{numeric: latitude coordinates of  where data are collected.}

\item{lon}{numeric: longitude coordinates of  where data are collected.}

\item{elevation}{numeric: Elevation of the Default is 40.}

\item{timeformat}{numeric: Native period of aggregation. Default is "daily".}

\item{sourcedata}{Matrix or data.frame or ascii file of raw data.}

\item{field_delimiter}{char string: field delimiter of file. Default is comma ",".}

\item{timeseries}{xts R object Timeseries of data}

\item{CRS}{char string: Projection of coordinate in proj4 format.}

\item{data_maintaner}{char string: maintainer's  name or contact or its contact .}
}
\description{
\code{meteodata} Create meteodata object for rAedesSim.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso crisci \email{a.crisci@ibimet.cnr.it} ASL 2 LUCCA Marco Selmi \email{m.selmi@usl2.toscana.it}
}
\seealso{
\code{\link{},\link{biometeo}}
}
\keyword{data}
\keyword{metorological}

