% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mortality_rate.r
\name{mortality_rate}
\alias{mortality_rate}
\title{mortality_rate}
\usage{
mortality_rate(tair = 24, twater = 24, urel = 60, stage = "Eggs",
  air_dryness = FALSE)
}
\arguments{
\item{tair}{numeric:   Temperature in Celsius Degree of (I) air: for "Adult" stage - (II)  for "Diapausant Eggs" the Minimal winter seasonal temperature  .}

\item{twater}{numeric: Water Temperature in Celsius degree.}

\item{urel}{numeric: Relative humidity in \% ( 0-100).}

\item{stage}{char string: Name of stage ( Diapausant Eggs - Eggs - Larval - Pupae - Adult). Default is "Eggs".}

\item{air_dryness}{logical: if  the saturation deficit of air  is taken into account as supplementary agent for adult mortality.}
}
\value{
m_rate numeric: daily mortality rate for the stage.
}
\description{
Calculate daily mortality rates for Aedes in function to the stage considered.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso crisci \email{a.crisci@ibimet.cnr.it} ASL 2 LUCCA Marco Selmi \email{m.selmi@usl2.toscana.it}
}
\references{
Poletti et al,2011, Transmission Potential of Chikungunya Virus and Control Measures: The Case of Italy. PLoS ONE 6(5).
  Magori et al,2009, Skeeter Buster: A Stochastic, Spatially Explicit Modeling Tool for Studying Aedes aegypti Population Replacement and Population Suppression Strategies. PLoS Negl Trop Dis 3(9).
  Brady et al,2013, Modelling adult Aedes aegypti and Aedes albopictus survival at different temperatures in laboratory
  and field settings. Parasites & Vectors 2013, 6:351
}
\keyword{Mortality}
\keyword{rates}

