% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/psychro_function.r
\name{pvapsat}
\alias{pvapsat}
\title{pvapsat}
\usage{
pvapsat(tair = 24)
}
\arguments{
\item{t}{numeric: air temperature in Celsius}
}
\value{
numeric: saturation pressure of water (hPa)
}
\description{
Saturation pressure of water in air as function of temperature (°C)
}
\details{
The saturation pressure of water is the pressure at or above which water
would condense from from the air. It is a function of temperature only.
}

