% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tombino_tosc_wmodel.r
\docType{data}
\name{tombino_tosc_wmodel}
\alias{tombino_tosc_wmodel}
\title{tombino_tosc_wmodel}
\format{A lm R object.}
\source{
ASL 2 LUCCA Marco Selmi \email{marco.selmi@uslnordovest.toscana.it}
}
\usage{
tombino_tosc_wmodel
}
\description{
Multivariate linear model  of daily mean water temperature observed in a manhole in Lucca (Bacale locality lat=43.925696
lon=10.433191) based on  daily minimum and maximum air temperature and length of day.
}
\keyword{datasets}
