% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/trans_rates.r
\name{trans_rates}
\alias{trans_rates}
\title{trans_rates}
\usage{
trans_rates(tair = 24, twater = 24, flag_I = FALSE, inib = 0.63,
  transition = "Ovideposition")
}
\arguments{
\item{tair}{Numeric   Temperature in Celsius degree  for "Ovidepostion" transition.}

\item{twater}{Numeric Water Temperature in Celsius degree for  any transition except "Ovidepostion".}

\item{transition}{Character string. Name of transition :  "Ovidepostion" - "Eggs2larvae" - "Larvae2Pupae" - "Pupae2Adult". Default is "Ovidepostion".}
}
\value{
Daily rate (day^-1)  rate for the trasition state considered.
}
\description{
Calculate daily transition rates for Aedes in function to the stage considered.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso crisci \email{a.crisci@ibimet.cnr.it} ASL 2 LUCCA Marco Selmi \email{m.selmi@usl2.toscana.it}
}
\references{
Poletti P., Messeri G, Ajelli M., Vallorani R., Rizzo C., et al. (2011) Transmission Potential of Chikungunya Virus and Control Measures: The Case of Italy. PLoS ONE 6(5): e18860. doi:10.1371/journal.pone.0018860
           Magori K, Legros M, Puente ME, Focks DA, Scott TW, Lloyd AL, Gould F. 2009. Skeeter Buster: A Stochastic, Spatially Explicit Modeling Tool for Studying Aedes aegypti Population Replacement and Population Suppression Strategies. PLoS Negl Trop Dis 3(9): e508. doi:10.1371/journal.pntd.0000508
}
\keyword{Container}

