% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trappola_wmodel.r
\docType{data}
\name{trappola_wmodel}
\alias{trappola_wmodel}
\title{trappola_wmodel}
\format{A lm R object.}
\source{
Istituto di Biometeorologia Firenze Italy  Alfonso crisci \email{a.crisci@ibimet.cnr.it}
}
\usage{
trappola_wmodel
}
\description{
Multivariate linear model  of daily mean water temperature observed in a mosquito REDLAV trap in Italy based on  daily minimum and maximum air temperature and length of day.
}
\keyword{datasets}
