% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/viewwhere.r
\name{viewwhere}
\alias{viewwhere}
\title{viewwhere}
\usage{
viewwhere(object, zoom = 17)
}
\arguments{
\item{object}{rAedesSim object with location coordinate.}
}
\value{
Return leaflet map as view.
}
\description{
Visualize the location of object in a leaflet framework
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso crisci \email{a.crisci@ibimet.cnr.it} ASL 2 LUCCA Marco Selmi \email{m.selmi@usl2.toscana.it}
}
\references{
The function is a wrapper for leaflet R package RStudio \url{"http://rstudio.github.io/leaflet"}
}
\keyword{View}
\keyword{location}

