% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/weigthdry.r
\name{weigthdry}
\alias{weigthdry}
\title{weigthdry}
\usage{
weigthdry(drylength, tresh = 5)
}
\arguments{
\item{drylength}{numeric vector: length of dry spell as number of consecutive dry days.}

\item{tresh}{nuemric: threshold in days}
}
\value{
Numerical vector of cumulate sequence
}
\description{
Calculate a exponential weigth in function to the amunt of dry spell.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso crisci \email{a.crisci@ibimet.cnr.it} ASL 2 LUCCA Marco Selmi \email{m.selmi@usl2.toscana.it}
}
\seealso{
\code{\link{},\link{drydaycons}}
}
\keyword{Drought}
\keyword{severity}

