% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocontainer.r
\name{biocontainer}
\alias{biocontainer}
\title{biocontainer}
\usage{
biocontainer(type = "Trap  REDLAV ITA", nrecipients = 50,
  container_shape = 3, capacity = 1000, initVol = 750,
  frac_initVol = 0.75, surface_area = 50, underplate = FALSE,
  frac_underplate = 0, frac_covering = 0, lat = 43.5, lon = 11.27,
  elevation = 100, CRS = "+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs",
  sourcedata = NULL, meteodata = NULL, watermodel = NULL,
  watermodel_fit = FALSE, model_type = "gam", date_format = "YMD",
  field_delimiter = ",", ID = c("NA"), site_name = c("NA"))
}
\arguments{
\item{type}{character  Container type.  The default trap used in REDLAV project is indicated: volume 0.75 750 cm^3 lt, heigth 13 cm, Area 50 cm^2.Describe the tipology of cointainer considered. For example (i) cointaner (ii) manhole (iii) others.}

\item{nrecipients}{numeric  Represent the numerosity of container used in simulation and correspond  to BS the number of breeding sites . Default is 50.}

\item{container_shape}{numeric Numeric level indicates if the surface container the shape . Parallelepiped plate (1)  Cilynder plate(2) Unspecified (3). Default is 3.}

\item{capacity}{numeric  The maximum capacity of water in trap/container. Default ( 1 lt 1000 cm^3).}

\item{initVol}{numeric  The unitarian initial volume of water in trap/container. Default ( 0.75 lt 750 ml).}

\item{frac_initVol}{numeric Fraction of volume present in trap/location. Default is 0.}

\item{surface_area}{numeric The water-air surface of container in squared centimeter (cmq).}

\item{underplate}{logical  Presence of underplate. Default is FALSE.}

\item{frac_underplate}{numeric Fraction of underplate engaged ( 0-1) Default is 0.}

\item{frac_covering}{numeric Fraction of container covering  ( 0-1) Default is 0.}

\item{lat}{numeric Mean geographical latitude of container location. Default is 43.5.}

\item{lon}{numeric Mean geographical longitude of container location. Default is 11.27.}

\item{elevation}{numeric  Mean elevation  of container location. Default is 100.}

\item{CRS}{character Projection of coordinate in proj4 format.}

\item{sourcedata}{dataframe  Or ascii file consisting in   Water temperature measurements. Fields required are dates ( YYYY-MM-DD) format, Daily mean of water tmedwater.}

\item{meteodata}{object  rAedesim  object with the data useful for model water calibration.}

\item{watermodel}{logical Statistical model oject ( gam - lm - glm).Default is NULL.}

\item{watermodel_fit}{logical Fit a statistical model oject ( gam - lm - glm).Default is FALSE.}

\item{model_type}{character  Statistical model class.}

\item{date_format}{character Format of the dates in raw data. Default is YMD.}

\item{field_delimiter}{character field delimiter of file. Default is comma ",".}

\item{ID}{character ID of container's set.}

\item{site_name}{character   Name of place or location.}

\item{daily_evap_loss}{numeric Volume of water losses daily due to evaporation. Default is 0.}

\item{timeseries}{object xts R Timeseries of data}
}
\value{
biocontainer
}
\description{
biocontainer is the function able to instantiate a S3 object used in rAedesSim simulation. The one is referred to the main features of the container where eggs are layed and larvae/pupae growth.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso crisci \email{a.crisci@ibimet.cnr.it} ASL 2 LUCCA Marco Selmi \email{marco.selmi@uslnordovest.toscana.it }
}
\keyword{container}
