% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biometeo.r
\name{biometeo}
\alias{biometeo}
\title{biometeo}
\usage{
biometeo(meteo, i_biocontainer, startday = 1, deltatmax = 0,
  deltatmin = 0, deltatmed = 0, tresh_rain = 4, weigth_k = 5,
  CPPmin = 11.5, datemax = 110, bounddays = 10, varjd = 10,
  truncated = FALSE)
}
\arguments{
\item{startday}{integer Index of date when  simulation starts. Default is 1.}

\item{deltatmax}{numeric Mean Error Bias considered for maximum air temperature. Default is 0.}

\item{deltatmin}{numeric Mean Error Bias  considered for minimum air temperature. Default is 0.}

\item{deltatmed}{numeric Mean Error Bias  considered for mean air temperature. Default is 0.}

\item{weigth_k}{numeric Weighting Paraemter to take into account dry spell impact See also \code{weigthdry}. Default is 5.}

\item{datemax}{integer Julian day of maxim date of spring diapause awekaning . Default is 110 4th April.}

\item{bounddays}{numeric Days of incertainty for diapause emergency period bounds. Default is 10.}

\item{varjd}{numeric Days of variance for  diapause emergency courbes. Default is 10.}

\item{truncated}{logical Modeling with truncated gaussian respecting the limits. Default is FALSE.}

\item{meteodata}{object rAedesim meteodata object.}

\item{biocontainer}{object rAedesim biocontainer object}

\item{tresh_rain_dry}{numeric Rain threshold for effective precipitation. Default is 4.}
}
\value{
Return a biometeo object.
}
\description{
Biometeo is the function to instantiate a S3 object related to environmental parameter that to be used in biomodel function.
It require firstly a meteodata object and biocontainer object as arguments.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso crisci \email{a.crisci@ibimet.cnr.it} ASL LUCCA Marco Selmi \email{marco.selmi@uslnordovest.toscana.it }
}
\keyword{container}
