% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomodel.r
\name{biomodel}
\alias{biomodel}
\title{biomodel}
\usage{
biomodel(i_biometeo, i_biocontainer, i_biopopulation, i_bioparameters,
  state = c(L1 = 0, L2 = 0, L3 = 0, L4 = 0, L5 = 0, L6 = 0, L7 = 0, L8 = 0),
  time_interval = 24, stocastic = TRUE, n_sampling = 10,
  inibition = FALSE, ID_sim = "", saveparameter = FALSE)
}
\arguments{
\item{i_biometeo}{object rAedesSim object concerning biometeorological parameter of location where mosquito population are simulated.}

\item{i_biocontainer}{object rAedesSim object concerning trap or habitat considered.}

\item{i_biopopulation}{object rAedesSim object concerning mosquito population.}

\item{i_bioparameters}{object rAedesSim object concerning biological parameter of mosquito population.}

\item{state}{numeric The state vector used in time simulation.The values are number of individuals L1 : Ovideposition rate, L3 : Eggs Mortality rate, 
L4: Eggs2larvae transition rate, L5 : Pupae2Larvae transition rate, L6 Larvae Mortality,L7  Pupae Mortality
L8 Pupae2Adult transition rate, L10 Adult mortality Default: L1=0,L3=0,L4=0,L5=0,L6=0,L7=0,L8=0,L10=0}

\item{time_interval}{numeric Interval times of simulation.}

\item{stocastic}{logical performs stocastic elaboration by poisson sampling at the end of simulation for each state.}

\item{inibition}{logical taking into account density larvae inibition.Default is FALSE.}

\item{ID_sim}{character  ID of simulation.Default is FALSE.}

\item{saveparameter}{logical  Save results about paraemters. Default is FALSE.}

\item{n_sampling=10}{numeric : number of resampling.}
}
\value{
biomodel
}
\description{
The biomodel function is the core function that performs simulation.It consists in a S3 object where by using the others objects the simulation is performed.
             calculating the transitional rate of population by using daily environmental data. With these a a Runge-Kutta
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso crisci \email{a.crisci@ibimet.cnr.it} ASL LUCCA Marco Selmi \email{marco.selmi@uslnordovest.toscana.it }
}
\keyword{daily}
\keyword{modeling}
\keyword{simulation}
