% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioparameters.r
\name{bioparameters}
\alias{bioparameters}
\title{bioparameters}
\usage{
bioparameters(alfa_l = 1, alfa_a = 0, exp_lar = 1, l_density = 100,
  sex_ratio = 0.5, egn = 63, ef = 0.83, inib = 0, sspp = "Albopictus",
  genus_sspp = "Aedes", order_sspp = "Diptera", geo_area = "Tuscany",
  name_location = NA)
}
\arguments{
\item{numeric}{alfa_l Coefficient of competition between individuals at larval stage.}

\item{numeric}{alfa_a  Coefficient of competition between individuals at adult stage.}

\item{numeric}{exp_lar Exponential Gompertz parameter for competition modulation. Default=1.}

\item{numeric}{l_density Expected larval density. Default is 70.}

\item{numeric}{sexratio   Sex ratio of population. Default is 0.5.}

\item{numeric}{egn  Mean eggs . Default is 63.}

\item{numeric}{inib Inibition rate parameter ( 0-1) Default is 0.}

\item{string}{sspp Name of the species. Default is "Albopictus".}

\item{string}{genus_sspp  Name of the genus of the species. Default is "Aedes".}

\item{string}{order_sspp  Name of the order of the species. Default is "Diptera".}

\item{string}{geo_area  Name of geographic area. Default is "Tuscany".}

\item{string}{name_location  Name of location.}

\item{numeric}{name_location mean elevation  of container location. Default is 100.}
}
\value{
S3 object Bioparameters object
}
\description{
Bioparameters is a function to instantiate an S3 object  containing biological parameters used in simulation for rAedesSim .
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso crisci \email{a.crisci@ibimet.cnr.it} ASL 2 LUCCA Marco Selmi \email{m.selmi@usl2.toscana.it}
}
\keyword{biparameters}
