% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioparameters.r
\name{bioparameters}
\alias{bioparameters}
\title{bioparameters}
\usage{
bioparameters(alfa_l = 1, alfa_a = 0, exp_lar = 1, l_density = 100,
  sex_ratio = 0.5, egn = 16, ef = 0.83, inib = 0, sspp = "Albopictus",
  genus_sspp = "Aedes", order_sspp = "Diptera", geo_area = "Tuscany",
  name_location = NA)
}
\arguments{
\item{alfa_l}{numeric Coefficient of competition between individuals at larval stage.}

\item{alfa_a}{numeric  Coefficient of competition between individuals at adult stage.}

\item{exp_lar}{numeric Exponential Gompertz parameter for competition modulation. Default=1.}

\item{l_density}{numeric Expected larval density. Default is 70.}

\item{egn}{numeric the Mean eggs during ovideposition cycle. Default is 63.}

\item{inib}{numeric Inibition rate parameter ( 0-1) Default is 0.}

\item{sspp}{character Name of the species. Default is "Albopictus".}

\item{genus_sspp}{character  Name of the genus of the species. Default is "Aedes".}

\item{order_sspp}{character  Name of the order of the species. Default is "Diptera".}

\item{geo_area}{character Name of geographic area. Default is "Tuscany".}

\item{name_location}{character  Name of location.}

\item{sexratio}{numeric   Sex ratio of population. Default is 0.5.}

\item{name_location}{numeric mean elevation  of container location. Default is missing.}
}
\value{
S3 object Bioparameters object
}
\description{
Bioparameters is a function to instantiate an S3 object  containing biological parameters used in simulation for rAedesSim .
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso crisci \email{a.crisci@ibimet.cnr.it} ASL LUCCA Marco Selmi \email{marco.selmi@uslnordovest.toscana.it }
}
\keyword{bioparameters}
