% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biopopulation.r
\name{biopopulation}
\alias{biopopulation}
\title{biopopulation}
\usage{
biopopulation(eggs = 10000, larvae = 0, pupae = 0, adults = 0,
  eggs_diap = 10, ID = "NA")
}
\arguments{
\item{eggs}{numeric  Eggs Count.}

\item{larvae}{numeric Larvae Count.}

\item{pupae}{numeric Pupae Count.}

\item{adults}{numeric Mosquito Female Actual Count.}

\item{eggs_diap}{numeric Diapausant Eggs Count.}

\item{ID}{character ID label Population.}
}
\value{
object Return a rAedesim biopopulation object.
}
\description{
S3 class to describe biomodelday object to be used in rAedesim model.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso crisci \email{a.crisci@ibimet.cnr.it} ASL LUCCA Marco Selmi \email{marco.selmi@uslnordovest.toscana.it }
}
\keyword{container}
