% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biopopulation.r
\name{biopopulation}
\alias{biopopulation}
\title{biopopulation}
\usage{
biopopulation(eggs = 10000, larvae = 0, pupae = 0, adults = 0,
  eggs_diap = 10, ID = "NA")
}
\arguments{
\item{eggs}{Number  Eggs Count.}

\item{larvae}{Number Larvae Count.}

\item{pupae}{Number Pupae Count.}

\item{adults}{Number Mosquito Female Actual Count.}

\item{eggs_diap}{Number Diapausant Eggs Count.}

\item{ID}{character ID label Population.}
}
\value{
object Return a rAedesim biopopulation object.
}
\description{
S3 class to describe biomodelday object to be used in rAedesim model.
}
\author{
Alfonso crisci \email{a.crisci@ibimet.cnr.it} Istituto di Biometeorologia Firenze Italy  Marco Selmi  ASL 2 LUCCA  \email{m.selmi@usl2.toscana.it}
}
\keyword{container}
