% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/develop_rate.r
\name{develop_rate}
\alias{develop_rate}
\title{develop_rate}
\usage{
develop_rate(temp, DR25, EA, EI, TI, t25 = 298)
}
\arguments{
\item{temp}{numeric  Temperature in Celsius degree.}

\item{DR25}{numeric Developmental rate standard at 25 Celsius.}

\item{EA}{numeric Enthalpy of activation for determinate stage.}

\item{EI}{numeric Enthalpy of inactivation for determinate stage.}

\item{TI}{numeric Temperature of inactivation for determinate stage.}

\item{t25}{numeric Temperature of inactivation for determinate stage.}
}
\value{
rate
}
\description{
Calculate the developmental rates for Aedes spp. in function to the air temperature. The function is based and refered n Solari at al (2006) work.
}
\references{
Otero, M., Solari, H., Schweigmann, N., 2006. A stochastic population dynamic model for aedes aegypti: formulation and application to a city with temperate climate. Bull. Math. Biol. 68, 1945 974.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso crisci \email{a.crisci@ibimet.cnr.it} ASL 2 LUCCA Marco Selmi \email{m.selmi@usl2.toscana.it}
}
\keyword{develpmental}
\keyword{rate}
