% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drydaycons.r
\name{drydaycons}
\alias{drydaycons}
\title{drydaycons}
\usage{
drydaycons(prec, tresh = 1)
}
\arguments{
\item{prec}{numeric  vector of daily cumulated rainfall in mm.}

\item{tresh}{numeric Rainfall threshold in mm}
}
\value{
numeric Numerical vector of cumulate sequence
}
\description{
Calculate the sequence of consecutive dry days in function to a certain  treshshold.
}
\seealso{
\code{weigthdry}
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso crisci \email{a.crisci@ibimet.cnr.it} ASL 2 LUCCA Marco Selmi \email{marco.selmi@uslnordovest.toscana.it }
}
\keyword{days}
\keyword{dry}
