% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychro_function.r
\name{evaporation_params}
\alias{evaporation_params}
\title{evaporation_params}
\usage{
evaporation_params(tair = 24, twater = 21, urel = 60, v_air = 0.5,
  L = 0.1, A = NULL)
}
\arguments{
\item{tair}{numeric  Air temperature in degC.}

\item{twater}{numeric Water temperature in degC.}

\item{urel}{numeric  Air Relative umidity.}

\item{v_air}{numeric Velocity of air movement.}

\item{L}{numeric Dimension factor to caracterize the cointainer.A cyliner model is considered. Generally L is a diameter lenght measured in meter. Default is 0.1.}

\item{A}{numeric Area of air-water interface. Default is NULL.}
}
\value{
evaporation_params
}
\description{
The function return a list of parameters in order to assess the evaporative losses 
for a generic cylndric cointainer under forced convetion (v_air > 0.1). Parameters done are in order:the  Reynolds number,Raylegh number,
the Schmidt and Sherwood numbers,the hm coeffcient of vapor trasport, the mv evaporative loss in time in in kg/sec and mvday the evaporative loss in a day (mg/day).
}
