% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychro_function.r
\name{mass_transport_coef}
\alias{mass_transport_coef}
\title{mass_transport_coef}
\usage{
mass_transport_coef(tair = 23, twater = 21, urel = 50, v_air = 0.5)
}
\arguments{
\item{tair}{numeric Air temperature in degC.}

\item{twater}{numeric Water temperature in degC.}

\item{urel}{numeric Air Relative umidity.}

\item{v_air}{numeric Velocity of air movement (meter per second).}
}
\value{
mass_transport_coef
}
\description{
Mass transfer coefficient in function to air temperature in forced convection  (v_air > 0.1).
}
