% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteodata.r
\name{meteodata}
\alias{meteodata}
\title{meteodata}
\usage{
meteodata(station_name = "Pisa San Giusto",
  network = "Aeronautica Militare", data_type = "Simulation",
  standard = "SYREP", data_provider = "IBIMET CNR", data_maintainer = "",
  data_licence = "", date_format = "YMD", lat = 43, lon = 11,
  elevation = 40, timeformat = "daily", sourcedata = NULL,
  field_delimiter = ",", timezone = "Europe/Rome", timeseries = NULL,
  CRS = "+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs")
}
\arguments{
\item{station_name}{character name of station or location}

\item{network}{character name or project or network which data corresponds.}

\item{data_type}{character weather data class: forecast,observation, sensor micrometeorological data, physical parameter simulation}

\item{standard}{character weather data standard class. Default is SYREP.}

\item{data_provider}{character institution -Private data manager.}

\item{data_licence}{character Licence of data.}

\item{date_format}{character Format of the dates in raw data. Default is YYYY-MM-DD.}

\item{lat}{numeric  latitude coordinate.}

\item{lon}{numeric  longitude coordinate.}

\item{elevation}{numeric Elevation corresponding to data collected. Default is 40.}

\item{timeformat}{numeric Time period of data aggregation. Default is "Daily".}

\item{sourcedata}{matrix or data.frame or ascii file of raw data. Frame must be daily with at least 6 fields. dates ( in YYYY-MM-DD format ),tmed,tmax,tmin,rhum prec.}

\item{field_delimiter}{chararter field delimiter of file. Default is comma ",".}

\item{timezone}{sring Timezone. Default is "Europe/Rome".}

\item{timeseries}{object xts R Timeseries of data}

\item{CRS}{character Projection of coordinates in proj4 format.}

\item{data_maintaner}{character maintainer's  name or contact or its contact .}
}
\value{
meteodata
}
\description{
meteodata is the function to store raw weather information in a S3 object in order to performs the simulation.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso crisci \email{a.crisci@ibimet.cnr.it} ASL 2 LUCCA Marco Selmi \email{marco.selmi@uslnordovest.toscana.it }
}
\keyword{data}
\keyword{metorological}
