% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychro_function.r
\name{pvapsat}
\alias{pvapsat}
\title{pvapsat}
\usage{
pvapsat(tair = 24)
}
\arguments{
\item{t}{numeric air temperature in Celsius}
}
\value{
pvapsat
}
\description{
Saturation saturation pressure of water (hPa) as function of air temperature (degC).
}
\details{
The saturation pressure of water is the pressure at or above which water
would condense from from the air. It is a function of air temperature.
}
