% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trappola_wmodel.r
\docType{data}
\name{trappola_wmodel}
\alias{trappola_wmodel}
\title{trappola_wmodel}
\format{A lm R object.}
\source{
\url{http://www.ibimet.cnr.it/}
}
\usage{
trappola_wmodel
}
\description{
Numerical model based on water temperature observation in Lucca to asses this parameter from daily minimum and maximum temperature.
}
\keyword{datasets}
