% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewwhere.r
\name{viewwhere}
\alias{viewwhere}
\title{viewwhere}
\usage{
viewwhere(object, zoom = 17)
}
\arguments{
\item{object}{object rAedesSim object with location coordinate.}

\item{zoom}{numeric  map scale.}
}
\value{
Return a leaflet map as view.
}
\description{
Visualize the location of object in a leaflet framework
}
\references{
The function is a wrapper for leaflet R package RStudio \url{"http://rstudio.github.io/leaflet"}
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso crisci \email{a.crisci@ibimet.cnr.it} ASL 2 LUCCA Marco Selmi \email{m.selmi@usl2.toscana.it}
}
\keyword{View}
\keyword{location}
