% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weigthdry.r
\name{weigthdry}
\alias{weigthdry}
\title{weigthdry}
\usage{
weigthdry(drylength, tresh = 5)
}
\arguments{
\item{drylength}{numeric vector: length of dry spell as number of consecutive dry days.}

\item{tresh}{nuemric: threshold in days}
}
\value{
weigthdry
}
\description{
Calculate a exponential weigth to perform the impact of dry spell in function to its length. 
It is used to evaluate the impact of drought on mosquito population. A rain treshshold is considered. Default value is 5 mm.
}
\seealso{
\code{"drydaycons"}
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso crisci \email{a.crisci@ibimet.cnr.it} ASL 2 LUCCA Marco Selmi \email{m.selmi@usl2.toscana.it}
}
\keyword{drought,}
\keyword{dry}
\keyword{spell}
